Inline DoFnRunnerBase to its sole subclass
