+++ runners/core-java/src/test/java/org/apache/beam/runners/core/SimpleDoFnRunnerTest.java
@@ -36,7 +36,7 @@ import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
 /**
- * Tests for base {@link DoFnRunnerBase} functionality.
+ * Tests for {@link SimpleDoFnRunner} functionality.
  */
 @RunWith(JUnit4.class)
 public class SimpleDoFnRunnerTest {
